rename ReactionsController to ReactionsService.
